Add files via upload
This is a simple project where we create Decision Tree and visualize it graphically. The purpose is if we feed any new data to this classifier, it would be able to predict the right class accordingly. We create a model, predict it  and find the accuracy of the model of how well it has learnt.